requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM
openai>=1.0.0
anthropic>=0.18.0

# Image Processing
Pillow>=10.0.0
rembg>=2.0.50
opencv-python>=4.8.0

# Video Processing
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9

# Async & HTTP
aiohttp>=3.9.0
httpx>=0.25.0
tenacity>=8.2.0

# Utilities
loguru>=0.7.0
pyyaml>=6.0
tiktoken>=0.5.0

# Optional: Cloud Storage
# boto3>=1.28.0
# google-cloud-storage>=2.10.0
